+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Active
 
+### Added
+ - Interval tree from unanimity
+
 ## [1.8.0] - 2020-09-23
 
 ### Added